+++ screen_space/streamlines.py
@@ -251,5 +251,5 @@ def streamlines_to_strokes(
                 drawing_origin[0] + (x_coord := (p[0]*width_inv)) * drawing_x_axis[0] + (y_coord := (p[1]*height_inv)) * drawing_y_axis[0],
                 drawing_origin[1] + x_coord * drawing_x_axis[1] + y_coord * drawing_y_axis[1],
                 drawing_origin[2] + x_coord * drawing_x_axis[2] + y_coord * drawing_y_axis[2]
-            ) for p in sl])
+            ) for p in sl], dtype=np.float32)
         for sl in streamlines]
